requirements.txt
fastapi
uvicorn
deepgram-sdk==5.3.0
openai
spacy
python-dotenv
python-multipart
streamlit